Integrate server with database.
